permissions: move and rename PermissionPolicy input and output types
These types will soon be used in frontend permission clients as we
split out the queries that can return conditional decisions to protect
users from receiving them unexpectedly. This commit simply moves them
from permission-node to permission-common, with a small rename for
consistency.
