+++ .changeset/cuddly-cobras-grin.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-permission-common': patch
+---
+
+Add `PolicyQuery` and `PolicyDecision` types.
