+++ .changeset/healthy-trees-turn.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-permission-node': minor
+---
+
+Remove `PolicyAuthorizeQuery` and `PolicyDecision` types. Use `PolicyQuery` and `PolicyDecision` from @backstage/plugin-permission-common instead.
