+++ .changeset/soft-windows-dream.md
@@ -1,0 +1,6 @@
+---
+'@backstage/plugin-catalog-backend': patch
+'@backstage/plugin-permission-backend': patch
+---
+
+Use `PolicyQuery` and `PolicyDecision` types from @backstage/plugin-permission-common instead of @backstage/plugin-permission-node.
