+++ plugins/permission-common/src/types/api.ts
@@ -43,7 +43,7 @@ export enum AuthorizeResult {
 }
 
 /**
- * An individual authorization request for {@link PermissionClient#authorize}.
+ * An individual request for {@link PermissionClient#authorize}.
  * @public
  */
 export type AuthorizeQuery = {
@@ -58,6 +58,71 @@ export type AuthorizeQuery = {
 export type AuthorizeRequest = {
   items: Identified<AuthorizeQuery>[];
 };
+
+/**
+ * An request for a {@link @backstage/plugin-permission-node#PermissionPolicy}
+ * to evaluate a permission..
+ *
+ * @remarks
+ *
+ * This differs from {@link AuthorizeQuery} in that `resourceRef` should never
+ * be provided. This forces policies to be written in a way that's compatible
+ * with filtering collections of resources at data load time.
+ *
+ * @public
+ */
+export type PolicyQuery<TPermission extends Permission = Permission> = {
+  permission: TPermission;
+  resourceRef?: never;
+};
+
+/**
+ * A definitive decision returned by a
+ * {@link @backstage/plugin-permission-node#PermissionPolicy}.
+ *
+ * @remarks
+ *
+ * This indicates that the policy unconditionally allows (or denies) the
+ * request.
+ *
+ * @public
+ */
+export type DefinitivePolicyDecision = {
+  result: AuthorizeResult.ALLOW | AuthorizeResult.DENY;
+};
+
+/**
+ * A conditional decision returned by a
+ * {@link @backstage/plugin-permission-node#PermissionPolicy}.
+ *
+ * @remarks
+ *
+ * This indicates that the policy allows authorization for the request, given
+ * that the returned conditions hold when evaluated. The conditions will be
+ * evaluated by the corresponding plugin which knows about the referenced
+ * permission rules.
+ *
+ * Similar to {@link @backstage/permission-common#AuthorizeDecision}, but with
+ * the plugin and resource identifiers needed to evaluate the returned
+ * conditions.
+ * @public
+ */
+export type ConditionalPolicyDecision = {
+  result: AuthorizeResult.CONDITIONAL;
+  pluginId: string;
+  resourceType: string;
+  conditions: PermissionCriteria<PermissionCondition>;
+};
+
+/**
+ * A decision returned by a
+ * {@link @backstage/plugin-permission-node#PermissionPolicy}.
+ *
+ * @public
+ */
+export type PolicyDecision =
+  | DefinitivePolicyDecision
+  | ConditionalPolicyDecision;
 
 /**
  * A condition returned with a CONDITIONAL authorization response.
