+++ plugins/permission-node/src/policy/types.ts
@@ -15,67 +15,10 @@
  */
 
 import {
-  AuthorizeQuery,
-  AuthorizeResult,
-  PermissionCondition,
-  PermissionCriteria,
+  PolicyDecision,
+  PolicyQuery,
 } from '@backstage/plugin-permission-common';
 import { BackstageIdentityResponse } from '@backstage/plugin-auth-node';
-
-/**
- * An authorization request to be evaluated by the {@link PermissionPolicy}.
- *
- * @remarks
- *
- * This differs from {@link @backstage/permission-common#AuthorizeQuery} in that `resourceRef`
- * should never be provided. This forces policies to be written in a way that's compatible with
- * filtering collections of resources at data load time.
- *
- * @public
- */
-export type PolicyAuthorizeQuery = Omit<AuthorizeQuery, 'resourceRef'>;
-
-/**
- * A definitive result to an authorization request, returned by the {@link PermissionPolicy}.
- *
- * @remarks
- *
- * This indicates that the policy unconditionally allows (or denies) the request.
- *
- * @public
- */
-export type DefinitivePolicyDecision = {
-  result: AuthorizeResult.ALLOW | AuthorizeResult.DENY;
-};
-
-/**
- * A conditional result to an authorization request, returned by the {@link PermissionPolicy}.
- *
- * @remarks
- *
- * This indicates that the policy allows authorization for the request, given that the returned
- * conditions hold when evaluated. The conditions will be evaluated by the corresponding plugin
- * which knows about the referenced permission rules.
- *
- * Similar to {@link @backstage/permission-common#AuthorizeDecision}, but with the plugin and resource
- * identifiers needed to evaluate the returned conditions.
- * @public
- */
-export type ConditionalPolicyDecision = {
-  result: AuthorizeResult.CONDITIONAL;
-  pluginId: string;
-  resourceType: string;
-  conditions: PermissionCriteria<PermissionCondition>;
-};
-
-/**
- * The result of evaluating an authorization request with a {@link PermissionPolicy}.
- *
- * @public
- */
-export type PolicyDecision =
-  | DefinitivePolicyDecision
-  | ConditionalPolicyDecision;
 
 /**
  * A policy to evaluate authorization requests for any permissioned action performed in Backstage.
@@ -95,7 +38,7 @@ export type PolicyDecision =
  */
 export interface PermissionPolicy {
   handle(
-    request: PolicyAuthorizeQuery,
+    request: PolicyQuery,
     user?: BackstageIdentityResponse,
   ): Promise<PolicyDecision>;
 }
